chore: refactor code and add more docs
